fix(tests): correct test expectations to match actual behavior
- Fix GameNotFoundError to use ErrCodeGameNotFound instead of ErrCodeNotFound
- Fix RateLimitError test: expect window.String() not "Retry-After"
- Fix AppError_Error test: account for ErrCodeInternal default details
- Fix ErrorInfo_JsonMarshal test: JSON numbers unmarshal as float64
- Fix TestEnhancedRepo_BulkCreate: add required Sort/Order fields

+++ internal/errors/response_test.go
@@ -299,7 +299,7 @@ func TestErrorInfo_JsonMarshal(t *testing.T) {
 	assert.Equal(t, errorInfo.TraceID, unmarshaled.TraceID)
 	assert.Equal(t, errorInfo.Severity, unmarshaled.Severity)
 	assert.Equal(t, errorInfo.Retryable, unmarshaled.Retryable)
-	assert.Equal(t, 123, unmarshaled.Context["user_id"])
+	assert.Equal(t, float64(123), unmarshaled.Context["user_id"])
 }
 
 func TestAPIResponse_JsonMarshal(t *testing.T) {
